requirements.txt
argcomplete==3.0.8
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
djangorestframework==3.14.0
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
Pillow==9.5.0
placebo==0.9.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
urllib3==2.0.2
Werkzeug==2.3.4
zappa==0.57.0
